refactor: move custom_integrations to src/ folder
