+++ src/custom_integrations/tixier_mita_lab/raspberrypi_serve.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+while pwd | grep -q src
+do
+  cd ..
+done
+
+python3 -m src.custom_integrations.tixier_mita_lab.custom_raspberrypi_headless_zmq
